Cleanup and package restructure
