+++ src/viam/components/types.py
@@ -2,4 +2,4 @@ import warnings
 
 from viam.media.video import CameraMimeType, RawImage  # noqa: F401
 
-warnings.warn("The types module is deprecated. Please find the appropriate types in the media package", DeprecationWarning)
+warnings.warn("The viam.types module is deprecated. Please find the appropriate types in the media package", DeprecationWarning)
